Revise storage solution & Add Javascript Interfaces
